refactor: :art: changes after the review ( more to come leter )
renaming functions, implementing Chakra Theme, defining Tables Type externally, changing file structure, adding local board background graphic
