--- src/components/PlanRoom/RoomTables/ItemTypes.tsx
@@ -1,5 +1,0 @@
-const ItemTypes = {
-  TABLE: 'table',
-};
-
-export default ItemTypes;
