--- src/components/PlanRoom/SaveRoomTables/SaveBoard.tsx
@@ -1,7 +1,0 @@
-const SaveBoard = (tables: {
-  [key: string]: { top: number; left: number; text: string; width: number; height: number; borderRadius: number };
-}) => {
-  console.log(tables);
-};
-
-export default SaveBoard;
