+++ src/components/molecules/PlanRoom/RoomTables/TableDNDTypes/index.tsx
@@ -1,0 +1,5 @@
+const ItemTypes = {
+  TABLE: 'table',
+};
+
+export default ItemTypes;
