+++ src/components/molecules/PlanRoom/SaveRoomTables/SaveBoard/index.tsx
@@ -1,0 +1,7 @@
+import { TablesType } from 'components/molecules/PlanRoom/RoomTables/TableComponentsType';
+
+const SaveBoard = (tables: TablesType) => {
+  console.log(tables);
+};
+
+export default SaveBoard;
