+++ src/components/molecules/PlanRoom/Tittle/index.tsx
@@ -1,0 +1,9 @@
+import { Box } from '@chakra-ui/react';
+
+const PlanRoomTitle = () => (
+  <Box bg="white" h="4em" w="50%" px="2em" py="0.2em" color="black" fontWeight="bold">
+    Plan sali
+  </Box>
+);
+
+export default PlanRoomTitle;
