+++ src/components/molecules/PlanRoom/styled/breakpointsButton.tsx
@@ -1,0 +1,12 @@
+import { createBreakpoints } from '@chakra-ui/theme-tools';
+
+const breakpoints = createBreakpoints({
+  base: '100%',
+  sm: '100%',
+  md: '100%',
+  lg: '100%',
+  xl: '100%',
+  '2xl': '100%',
+});
+
+export default breakpoints;
