+++ src/components/molecules/PlanRoom/styled/breakpointsFontSize.tsx
@@ -1,0 +1,12 @@
+import { createBreakpoints } from '@chakra-ui/theme-tools';
+
+const breakpointsFontSize = createBreakpoints({
+  base: '60%',
+  sm: '70%',
+  md: '80%',
+  lg: '80%',
+  xl: '80%',
+  '2xl': '90%',
+});
+
+export default breakpointsFontSize;
