+++ src/components/molecules/Supplier/AddSupplierTitle/index.tsx
@@ -1,0 +1,9 @@
+import { Box } from '@chakra-ui/react';
+
+const AddSupplierTitle = () => (
+  <Box bg="white" h="4em" w="50%" px="2em" py="0.2em" color="black" fontWeight="bold">
+    Dodawanie dostawcy
+  </Box>
+);
+
+export default AddSupplierTitle;
