--- src/styled/PlanRoom/breakpointsFontSize.tsx
@@ -1,12 +1,0 @@
-import { createBreakpoints } from '@chakra-ui/theme-tools';
-
-const breakpointsFontSize = createBreakpoints({
-  base: '60%',
-  sm: '70%',
-  md: '80%',
-  lg: '80%',
-  xl: '80%',
-  '2xl': '90%',
-});
-
-export default breakpointsFontSize;
