--- src/styled/Supplier/AppSupplierData/breakpoints.tsx
@@ -1,11 +1,0 @@
-import { createBreakpoints } from '@chakra-ui/theme-tools';
-
-const breakpoints = createBreakpoints({
-  sm: '100%',
-  md: '100%',
-  lg: '48%',
-  xl: '48%',
-  '2xl': '48%',
-});
-
-export default breakpoints;
