--- src/styled/Supplier/AppSupplierData/breakpointsMargin.tsx
@@ -1,12 +1,0 @@
-import { createBreakpoints } from '@chakra-ui/theme-tools';
-
-const breakpointsMargin = createBreakpoints({
-  base: '-6%',
-  sm: '-6%',
-  md: '-4%',
-  lg: '-6%',
-  xl: '-6%',
-  '2xl': '-6%',
-});
-
-export default breakpointsMargin;
